Refactored automation, added a couple parsers:
Automation was changed as followed:
- before, special cases (repos with multiple grammars) were handled in
  the Go code, by adding a function to handle that case;
- that was changed by making the Grammar format more flexible, allowing
  it to pass in files from anywhere in the repo as input, as well as
  adding the ability to set the source root.

Concretely: usually, parsers have a src folder in the root, and then the
src/parser.c, src/tress_sitter/parser.h files (and possible scanner).

However, in case of repos that hold multiple grammars, such as
typescript, the have the following structure: gramma1/src/...,
grammarN/src/...

By introduction of `SrcRoot` field, now we can express in grammars.json
both repos with simple structure as well as repos with multiple grammars
structure. This makes it far simpler to add the latter in the future and
also makes it more clear, as now all the grammars that are exposed by a
repo are present in grammars.json and PARSERS.md.

Also, now the repo has a flat structure in all cases (i.e., normally we
had /ada, /java, /json, etc. but for repos with multiple grammars we had
/xml/xml, /xml/dtd, /typescript/typescript, /typescript/tsx, etc.).

Now each language we have in grammar.json/PARSERS.md has a corresponding
top level folder in the repo.

Tested the new automation right away of course :-)

- updated Python;
- added ocaml_interface;
- added php_only;
- documented dtd and tsx in PARSERS (they were actually already supported
  but not visible in PARSERS, as per the above explanation).